requirements.txt
flask==2.3.2
flask-cors==4.0.0
openai-whisper==20231117
torch>=1.9.0
torchaudio>=0.9.0
transformers>=4.21.0
requests>=2.28.0
psutil>=5.9.0
werkzeug>=2.3.0
jinja2>=3.1.0
gunicorn>=20.1.0 